% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{Arc}
\alias{Arc}
\title{specify the properties of an arc in a flowchart}
\usage{
Arc(from, to, rate, label = NULL, edge = NULL)
}
\arguments{
\item{from}{the name of the compartment that this flow originates.}

\item{to}{the name of the compartment that this flow goes to.}

\item{rate}{the rate of the flow, i.e., the label. An R expression}

\item{label}{the property of the label, speficied by the Label function.}

\item{edge}{ghe property for the arc, such as "bend left" (or "(") or "bend right" (or ")").}
}
\value{
a character holding the tikz command for the property of the label
}
\description{
An arc represents a transition.
}
