% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\name{attach.function}
\alias{attach.function}
\title{attach a function to be used by the models.}
\usage{
attach.function(..., functions = NULL)
}
\arguments{
\item{...}{named arguments, each specifies a function}

\item{functions}{a named list of functions, where the name are the functions used
in models}
}
\description{
attach a function to be used by the models.
}
\details{
a NULL value removes a function from the list of attached functions.
}
