% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Beta}
\alias{Beta}
\title{The beta distribution}
\usage{
Beta(shape1, shape2)
}
\arguments{
\item{shape1}{Optional first shape parameter.}

\item{shape2}{Optional second shape parameter.}
}
\value{
A \code{Distribution} object.
}
\description{
The beta distribution.
}
