% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibrator.R
\docType{class}
\name{Calibrator}
\alias{Calibrator}
\title{The base R6 class for calibrators}
\description{
The base R6 class for calibrators

The base R6 class for calibrators
}
\details{
A calibrator calibrates the model to a dataset. Like a simulator, 
to calibrate the model to a dataset, a calibrator
expects the known and unknown initial conditions and the parameter values, 
and the data to calibrate to. In addition, it also expects a mapping from 
model solutions to observation variables.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{the names of all parameters}

\item{\code{details}}{the details of the fitting output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Calibrator-new}{\code{Calibrator$new()}}
\item \href{#method-Calibrator-calibrate}{\code{Calibrator$calibrate()}}
\item \href{#method-Calibrator-clone}{\code{Calibrator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calibrator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Calibrator-new}{}}}
\subsection{Method \code{new()}}{
initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calibrator$new(
  model,
  time,
  data,
  ...,
  cumulative = FALSE,
  mapping = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to calibrate}

\item{\code{time}}{either a numeric vector containing the times (including the 
initial time) of the ODE solution that corresponds to the data, or a 
character value giving the name of the column in data that corresponds 
to time.}

\item{\code{data}}{a data.frame object containign the data for the calibration}

\item{\code{...}}{each argument is a formula defining the maps between 
the data columns and the model variables. Please see the details section.}

\item{\code{cumulative}}{whether the data is cumulative}

\item{\code{mapping}}{a named list specifying the mapping from data columns
to quantities in the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A mapping is a named argument, where name is the
data colummn name, and value corresponds to the model variables (or an 
expression to calculate from the model variables.)
Calibrate the model to data
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calibrator-calibrate"></a>}}
\if{latex}{\out{\hypertarget{method-Calibrator-calibrate}{}}}
\subsection{Method \code{calibrate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calibrator$calibrate(initial.values, parms, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial.values}}{the initial values for the model. The parameters 
that need to be estimate should be NA, those that do not need to be
estimated must contain a finite value.}

\item{\code{parms}}{the parameter values of the model. The parameters 
that need to be estimate should be NA, those that do not need to be
estimated must contain a finite value.}

\item{\code{...}}{extra arguments to be passed to calibrators}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Calibrator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Calibrator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Calibrator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
