% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{change.extension}
\alias{change.extension}
\title{change the extension of a file}
\usage{
change.extension(file, ext.from, ext.to)
}
\arguments{
\item{file}{a character giving the file path}

\item{ext.from}{the extension to change from}

\item{ext.to}{the extension to change to}
}
\description{
change the extension of a file
}
\details{
if the file has the extension given by ext.from, then 
the extension is changed. Otherwise, the extension given by ext.to
is appended to file.
}
\examples{
# both of the following gives "test.pdf"
change.extension("test.tex", ".tex", ".pdf")
change.extension("test", ".tex", ".pdf")
}
