% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compartmental.R
\docType{class}
\name{Compartmental}
\alias{Compartmental}
\title{An R6 class representing a compartmental model}
\description{
An R6 class representing a compartmental model

An R6 class representing a compartmental model
}
\details{
This is a subclass of Model. Unlike a Model object which gives the
ODE for the rate of change of the states, a Compartmental model 
describes the model using compartments and transitions between 
compartments. Each transition may correspond to an event. The ODE 
system can then be derived by summing the transitions that flow from
the compartment (negative flows) and those that flow to the compartment
(positive flows). Because a compartmental model can describe events,
they can also be used for stochastic simulations (such as the Gillespie
method).

The deleted compartment or substitution is converted to a 
parameter. A parameter cannot be deleted. If the equation is changed
so that a parameter is not used by the model any more, then the paramter
is automatically removed.

The formula can be specified by a formula, which has the form
from -> to ~ rate or to <- from ~ rate
where the rate can be either an expression, or percapita(expression). 
Alternatively the from, to, or rate (and percapita) can be specified by the
arguments of the same name. 

A transition must be named. If the name is not provided, one is 
automatically generated. Any further change to the transition must be
referred by the name.

The from or to can be null. If from is NULL, then the transition is
an input. If to is NULL, then the transition is an output. If both 
the name is deleted.

If the transition with the given name exists, then the transition is changed.
}
\examples{

## ------------------------------------------------
## Method `Compartmental$new`
## ------------------------------------------------

# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
print(SIR)

## ------------------------------------------------
## Method `Compartmental$compartment`
## ------------------------------------------------

# an SIR model
SIR = Compartmental$new()
SIR$compartment("S")$compartment("I")$compartment(R)
SIR$transition(S->I ~ beta*S*I/N, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
SIR$where(N=S+I+R)
print(SIR)

## ------------------------------------------------
## Method `Compartmental$delete`
## ------------------------------------------------

# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
print(SIR)

## ------------------------------------------------
## Method `Compartmental$transition`
## ------------------------------------------------

# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
SIR$transition(S->R ~ v, percapita=TRUE, name="vaccination")
# the following command changes the "vaccination" transition
SIR$transition(S->NULL ~ v*S, name="vaccination")
# this following command delete the "vaccination" transition, because
# the rate is default to NULL.
SIR$transition(name="vaccination")
SIR$transition(NULL->S ~ lambda, name="births")
SIR$transition(S -> NULL ~ percapita(mu), name="death.S")
SIR$transition(I -> NULL ~ percapita(mu), name="death.I")
SIR$transition(R -> NULL ~ percapita(mu), name="death.R")
print(SIR)
}
\section{Super class}{
\code{\link[REpiSim:Model]{REpiSim::Model}} -> \code{Compartmental}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{transitions}}{a read-only field to access the transitions.}

\item{\code{representation}}{a read-only active field that returns the representation of the model
it returns a list that contains self$compartments, self$transitions and
self$substitutions. The compartmental model can then be reconstructed from the representation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Compartmental-new}{\code{Compartmental$new()}}
\item \href{#method-Compartmental-compartment}{\code{Compartmental$compartment()}}
\item \href{#method-Compartmental-delete}{\code{Compartmental$delete()}}
\item \href{#method-Compartmental-transition}{\code{Compartmental$transition()}}
\item \href{#method-Compartmental-format}{\code{Compartmental$format()}}
\item \href{#method-Compartmental-clone}{\code{Compartmental$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Model" data-id="rename"><a href='../../REpiSim/html/Model.html#method-Model-rename'><code>REpiSim::Model$rename()</code></a></span></li>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Model" data-id="where"><a href='../../REpiSim/html/Model.html#method-Model-where'><code>REpiSim::Model$where()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-new"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$new(..., t = "t", file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the names of the compartment, or substitutions}

\item{\code{t}}{the name of the independent variable, either a name or a string}

\item{\code{file}}{if not NULL, a path or connection to a model file 
to read the model from}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
print(SIR)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-compartment"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-compartment}{}}}
\subsection{Method \code{compartment()}}{
define a compartment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$compartment(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the compartment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an invisible self for chaining methods
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# an SIR model
SIR = Compartmental$new()
SIR$compartment("S")$compartment("I")$compartment(R)
SIR$transition(S->I ~ beta*S*I/N, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
SIR$where(N=S+I+R)
print(SIR)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-delete"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-delete}{}}}
\subsection{Method \code{delete()}}{
delete a compartment, a substitution or a transition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$delete(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the compartment, substitution or transition
to delete.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an invisible self for chaining methods
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
print(SIR)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-transition"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-transition}{}}}
\subsection{Method \code{transition()}}{
define or change a transition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$transition(
  formula = NULL,
  ...,
  from = NULL,
  to = NULL,
  rate = NULL,
  percapita = FALSE,
  name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{the transition can be fullt specified by a formula. Please see the
details.}

\item{\code{...}}{named arguments specifying substitutions for the parameters
used by this transition.}

\item{\code{from}}{the name of the compartment that the transition originates}

\item{\code{to}}{the name of the compartment that the transition flows into}

\item{\code{rate}}{the rate of the transition, an expression. If rate is NULL, the transition is deleted.}

\item{\code{percapita}}{whether the rate is per capita, i.e., the total rate would 
be the per capita rate multiplied by the from compartment.}

\item{\code{name}}{the name of the transition, a character. If NULL, the name
is automatically generated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the name of the transition
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# an SIR model
SIR = Compartmental$new(S, I, R)
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
SIR$transition(S->R ~ v, percapita=TRUE, name="vaccination")
# the following command changes the "vaccination" transition
SIR$transition(S->NULL ~ v*S, name="vaccination")
# this following command delete the "vaccination" transition, because
# the rate is default to NULL.
SIR$transition(name="vaccination")
SIR$transition(NULL->S ~ lambda, name="births")
SIR$transition(S -> NULL ~ percapita(mu), name="death.S")
SIR$transition(I -> NULL ~ percapita(mu), name="death.I")
SIR$transition(R -> NULL ~ percapita(mu), name="death.R")
print(SIR)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-format"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-format}{}}}
\subsection{Method \code{format()}}{
format the class for printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Compartmental-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Compartmental-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Compartmental$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
