% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE.R
\docType{class}
\name{Equilibrium}
\alias{Equilibrium}
\title{An R6 class to calculate the equilibrium of a model}
\description{
An R6 class to calculate the equilibrium of a model

An R6 class to calculate the equilibrium of a model
}
\details{
Calculate the equilibrium of the model with a given initial condition
}
\examples{
# an SIR model
SIR = Model$new(title="SIR")
SIR$compartment(S ~ -beta*S*I/N)$
  compartment(I ~ beta*S*I/N - gamma*I)$
  compartment(R ~ gamma*I)$
  where(N = S + I + R)
eq = Equilibrium$new(SIR, vary="beta", range=seq(0, 0.4, by=0.01))
ode$simulate(
  time=0:40, 
  y0=c(S=10000, I=1, R=0), 
  parms=c(beta=0.4, gamma=0.2), 
  alias = FALSE) # N is not returned.

}
\section{Super classes}{
\code{\link[REpiSim:Simulator]{REpiSim::Simulator}} -> \code{\link[REpiSim:ODE]{REpiSim::ODE}} -> \code{Equilibrium}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Equilibrium-new}{\code{Equilibrium$new()}}
\item \href{#method-Equilibrium-simulate}{\code{Equilibrium$simulate()}}
\item \href{#method-Equilibrium-clone}{\code{Equilibrium$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Equilibrium-new"></a>}}
\if{latex}{\out{\hypertarget{method-Equilibrium-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Equilibrium$new(
  model,
  vary = NULL,
  range = NULL,
  max.iter = 100,
  error = 1e-06
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to simulate. It must be an object of a subclass
of `Model`.}

\item{\code{vary}}{the name of a parameter to value, a character}

\item{\code{range}}{the range of the varying parameter, a numerical vector}

\item{\code{max.iter}}{the maximum iterations to search for the equilibrium}

\item{\code{error}}{a numeric value giving the 2-norm of the tolerance}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Equilibrium-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Equilibrium-simulate}{}}}
\subsection{Method \code{simulate()}}{
simulate the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Equilibrium$simulate(t, y0, parms = NULL, vars = names(y0), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{a numeric vector of times.}

\item{\code{y0}}{a numeric vector of initial conditions.}

\item{\code{parms}}{a numeric vector of parameter values.}

\item{\code{vars}}{a character vector specifying the variable names to be returned. 
A variable can be either a compartment or a substitution}

\item{\code{...}}{extra arguments to be passed to the `ode` method
substitutions that depend on the compartments are calculated 
and returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame which rows correspond to each time in `t`, 
and columns correspond to the variables specified in vars.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Equilibrium-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Equilibrium-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Equilibrium$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
