% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Exponential}
\alias{Exponential}
\title{The exponential distribution}
\usage{
Exponential(rate)
}
\arguments{
\item{rate}{Optional fixed rate. If omitted, it is computed from the model mean.}
}
\value{
A \code{Distribution} object.
}
\description{
The exponential distribution.
}
