% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{Expression}
\alias{Expression}
\title{constructor}
\description{
constructor

constructor
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parms}}{the parameters used int he expression}

\item{\code{functions}}{the function named used int he expression}

\item{\code{expr}}{the original expression}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Expression-new}{\code{Expression$new()}}
\item \href{#method-Expression-substitute}{\code{Expression$substitute()}}
\item \href{#method-Expression-clone}{\code{Expression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Expression-new"></a>}}
\if{latex}{\out{\hypertarget{method-Expression-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Expression$new(expr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{the expression to parse
Perform a substitute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Expression-substitute"></a>}}
\if{latex}{\out{\hypertarget{method-Expression-substitute}{}}}
\subsection{Method \code{substitute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Expression$substitute(subs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subs}}{a named list of substitutions}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
For each name in the list, the variable with that name in the
expression is replaced by the corresponding value in the list.
}

\subsection{Returns}{
an expression with the substitution
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Expression-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Expression-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Expression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
