% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{flowchart}
\alias{flowchart}
\title{generate the flow chart for a compartmental model}
\usage{
flowchart(model, tex, ..., preambles = NULL)
}
\arguments{
\item{model}{an object of Compartmental class}

\item{tex}{an object of TexFormatter class}

\item{...}{the tikz commands. Compartment descriptions and 
transition descriptions must be given by named arguments, which names
must be the name for the compartment or transition, and the value is 
a description given by Node and Arc functions.}

\item{preambles}{a character holding a sequence of latex commands as latex preambles.}
}
\value{
a character holding a tikzpicture latex environment.
}
\description{
generate the flow chart for a compartmental model
}
\details{
When calling the Node function, the name and label of the compartment
does not need to be passed. The name it is given as the name of the argument.
The label is generated using tex. Similarly, when calling Arc, 
the from, to and rate need not be specified.
}
\examples{
# an SIR model
SIR = Compartmental$new(S, I, R, title="SIR")
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
flowchart(SIR, TexFormatter$new(),
  S=Node(at=c(0,0)),
  I=Node(at=c(2,0)),
  R=Node(at=c(4,0))
)
}
