% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Gamma}
\alias{Gamma}
\title{The gamma distribution}
\usage{
Gamma(shape, scale)
}
\arguments{
\item{shape}{Optional shape parameter.}

\item{scale}{Optional scale parameter.}
}
\value{
A \code{Distribution} object.
}
\description{
The gamma distribution.
}
