% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{\%*\%}
\alias{\%*\%}
\title{Define a generic multiplication operator}
\usage{
a \%*\% b
}
\arguments{
\item{a}{the first argument}

\item{b}{the second argument}
}
\value{
the product of a and b
}
\description{
Define a generic multiplication operator
}
