% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{Label}
\alias{Label}
\title{specify the properties of a label in a flowchart}
\usage{
Label(pos = NULL, direction = Label.directions)
}
\arguments{
\item{pos}{a value in the interval [0,1] specifying the position
of the label on the arc, 0 is at the starting point, 1 is at the
end point. Default to 0.5.}

\item{direction}{the relative position of the label to the arc}
}
\value{
a character holding the tikz command for the property of the label
}
\description{
A label is the rate of a transition.
}
