% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeastSquare.R
\docType{class}
\name{LeastSquare}
\alias{LeastSquare}
\title{The calibrator using the least square method.}
\description{
The calibrator using the least square method.

The calibrator using the least square method.
}
\section{Super classes}{
\code{\link[REpiSim:Calibrator]{REpiSim::Calibrator}} -> \code{\link[REpiSim:Optimizer]{REpiSim::Optimizer}} -> \code{LeastSquare}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LeastSquare-new}{\code{LeastSquare$new()}}
\item \href{#method-LeastSquare-clone}{\code{LeastSquare$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Optimizer" data-id="calibrate"><a href='../../REpiSim/html/Optimizer.html#method-Optimizer-calibrate'><code>REpiSim::Optimizer$calibrate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeastSquare-new"></a>}}
\if{latex}{\out{\hypertarget{method-LeastSquare-new}{}}}
\subsection{Method \code{new()}}{
initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeastSquare$new(
  model,
  time,
  data,
  ...,
  log = FALSE,
  cumulative = FALSE,
  mapping = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to calibrate}

\item{\code{time}}{either a numeric vector containing the times (including the 
initial time) of the ODE solution that corresponds to the data, or a 
character value giving the name of the column in data that corresponds 
to time.}

\item{\code{data}}{a data.frame object containign the data for the calibration}

\item{\code{...}}{each argument is a formula defining the maps between 
the data columns and the model variables. Please see the details section.}

\item{\code{log}}{boolean indicating whether to fit in log scale}

\item{\code{cumulative}}{whether the data is cumulative}

\item{\code{mapping}}{a list specifying the mapping from data columns to model variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A mapping is a named argument, where name is the
data colummn name, and value corresponds to the model variables (or an 
expression to calculate from the model variables.)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LeastSquare-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LeastSquare-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LeastSquare$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
