% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{Likelihood}
\alias{Likelihood}
\title{An R6 class that defines a likelihood function to be used by a Calibrator object}
\description{
An R6 class that defines a likelihood function to be used by a Calibrator object

An R6 class that defines a likelihood function to be used by a Calibrator object
}
\details{
Each likelihood object must provide a read-only active field
name par specifying likelihood parameters, and function logL for log
likelihood. These must be implemented by subclasses
The log likelihood function.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{par}}{return a vector of parameter names for this likelihood 
distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Likelihood-logL}{\code{Likelihood$logL()}}
\item \href{#method-Likelihood-clone}{\code{Likelihood$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Likelihood-logL"></a>}}
\if{latex}{\out{\hypertarget{method-Likelihood-logL}{}}}
\subsection{Method \code{logL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood$logL(x, mean, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the observation to calculate the likelihood}

\item{\code{mean}}{the mean of the likelihood distribution}

\item{\code{...}}{other likelihood parameter values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a numeric value giving the likelihood.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Likelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Likelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
