% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexFormatter.R
\name{merge.list}
\alias{merge.list}
\title{A S3 method to merge two lists}
\usage{
\method{merge}{list}(x, y, ...)
}
\arguments{
\item{x}{a list}

\item{y}{a list}

\item{...}{unused}
}
\value{
a merged list
}
\description{
A S3 method to merge two lists
}
