% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metrop.R
\name{Metrop}
\alias{Metrop}
\title{A subclass of Baysian calibrator that uses the mcmc::metrop function}
\description{
A subclass of Baysian calibrator that uses the mcmc::metrop function

A subclass of Baysian calibrator that uses the mcmc::metrop function
}
\section{Super classes}{
\code{REpiSim::Calibrator} -> \code{REpiSim::Baysian} -> \code{Metrop}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{the}}{names of all parameters}

\item{\code{The}}{samples of the posterior distribution}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{the}}{names of all parameters}

\item{\code{The}}{samples of the posterior distribution}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Metrop-new}{\code{Metrop$new()}}
\item \href{#method-Metrop-continue}{\code{Metrop$continue()}}
\item \href{#method-Metrop-clone}{\code{Metrop$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Baysian" data-id="calibrate"><a href='../../REpiSim/html/Baysian.html#method-Baysian-calibrate'><code>REpiSim::Baysian$calibrate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrop-new"></a>}}
\if{latex}{\out{\hypertarget{method-Metrop-new}{}}}
\subsection{Method \code{new()}}{
initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrop$new(
  model,
  time,
  data,
  likelihood,
  ...,
  cumulative = FALSE,
  mapping = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to calibrate}

\item{\code{time}}{either a numeric vector containing the times (including the 
initial time) of the ODE solution that corresponds to the data, or a 
character value giving the name of the column in data that corresponds 
to time.}

\item{\code{data}}{a data.frame object containign the data for the calibration}

\item{\code{likelihood}}{a Distribution object specify the type of likelihood}

\item{\code{...}}{each argument is a formula defining the maps between 
the data columns and the model variables. Please see the details section.}

\item{\code{cumulative}}{whether the data is cumulative}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A mapping is a named argument, where name is the
data colummn name, and value corresponds to the model variables (or an 
expression to calculate from the model variables.)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrop-continue"></a>}}
\if{latex}{\out{\hypertarget{method-Metrop-continue}{}}}
\subsection{Method \code{continue()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrop$continue(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Metrop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Metrop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Metrop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
