% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\docType{class}
\name{MLE}
\alias{MLE}
\title{A maximum likelihood calibrator using the bbmle package}
\description{
A maximum likelihood calibrator using the bbmle package

A maximum likelihood calibrator using the bbmle package
}
\section{Super class}{
\code{\link[REpiSim:Calibrator]{REpiSim::Calibrator}} -> \code{MLE}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{parameters}}{the names of all parameters
field CI whether to calculate the CI using likelihood profiling}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MLE-new}{\code{MLE$new()}}
\item \href{#method-MLE-clone}{\code{MLE$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Calibrator" data-id="calibrate"><a href='../../REpiSim/html/Calibrator.html#method-Calibrator-calibrate'><code>REpiSim::Calibrator$calibrate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLE-new"></a>}}
\if{latex}{\out{\hypertarget{method-MLE-new}{}}}
\subsection{Method \code{new()}}{
initializer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLE$new(
  model,
  time,
  data,
  likelihood,
  ...,
  cumulative = FALSE,
  mapping = character(),
  CI = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to calibrate}

\item{\code{time}}{either a numeric vector containing the times (including the 
initial time) of the ODE solution that corresponds to the data, or a 
character value giving the name of the column in data that corresponds 
to time.}

\item{\code{data}}{a data.frame object containign the data for the calibration}

\item{\code{likelihood}}{a Lieklihood object specify the type of likelihood}

\item{\code{...}}{each argument is a formula defining the maps between 
the data columns and the model variables. Please see the details section.}

\item{\code{cumulative}}{whether the data is cumulative}

\item{\code{mapping}}{a named list specifying the mapping from data columns}

\item{\code{CI}}{a boolean indicating whether to calculate the CI using likelihood profiling}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A mapping is a named argument, where name is the
data colummn name, and value corresponds to the model variables (or an 
expression to calculate from the model variables.)
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MLE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MLE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
