% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{NBinom}
\alias{NBinom}
\title{The negative binomial distribution}
\usage{
NBinom(size, prob)
}
\arguments{
\item{size}{Optional size parameter.}

\item{prob}{Optional probability parameter.}
}
\value{
A \code{Distribution} object.
}
\description{
The negative binomial distribution.
}
\details{
If one of \code{size} or \code{prob} is omitted, it is computed from the model
mean and the other parameter. If both are omitted, \code{prob} is treated as a
free likelihood parameter and \code{size} is computed from \code{mean} and
\code{prob}.
}
