% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{Node}
\alias{Node}
\title{declare a node for a flowchart}
\usage{
Node(
  name,
  label,
  at,
  shape = Node.shapes,
  rounded.corners = TRUE,
  fill = NULL,
  ...
)
}
\arguments{
\item{name}{the name of the compartment}

\item{label}{the latex label for the compartment}

\item{at}{a numeric vector giving the coordinate of the node}

\item{shape}{a character giving the shape of the compartment}

\item{rounded.corners}{if the shape should use rounded corners}

\item{fill}{an R color for filling the shape}

\item{...}{other properties for the node}
}
\value{
a character holding the tikz command for a node
}
\description{
A node in a flowchart represents a compartment.
}
