% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Normal}
\alias{Normal}
\title{The normal distribution}
\usage{
Normal(mean, sd)
}
\arguments{
\item{mean}{Optional fixed mean. If omitted, it is computed from the model mean.}

\item{sd}{Optional fixed standard deviation. If omitted, \code{sd} is a free
likelihood parameter.}
}
\value{
A \code{Distribution} object.
}
\description{
The normal distribution.
}
