% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\docType{class}
\name{Optimizer}
\alias{Optimizer}
\title{A subclass of Calibrator that uses an optimizer and an initial guess.}
\description{
A subclass of Calibrator that uses an optimizer and an initial guess.

A subclass of Calibrator that uses an optimizer and an initial guess.
}
\section{Super class}{
\code{\link[REpiSim:Calibrator]{REpiSim::Calibrator}} -> \code{Optimizer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Optimizer-clone}{\code{Optimizer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Calibrator" data-id="calibrate"><a href='../../REpiSim/html/Calibrator.html#method-Calibrator-calibrate'><code>REpiSim::Calibrator$calibrate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Calibrator" data-id="initialize"><a href='../../REpiSim/html/Calibrator.html#method-Calibrator-initialize'><code>REpiSim::Calibrator$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
