% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{order}
\alias{order}
\title{return the order of expressions by dependence.}
\usage{
order(exprs)
}
\arguments{
\item{exprs}{a named list of expressions}
}
\value{
a character vector giving the order of expressions
}
\description{
return the order of expressions by dependence.
}
\details{
An expression may need the value of other expressions to be evaluated.
}
