% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Poisson}
\alias{Poisson}
\title{The Poisson distribution}
\usage{
Poisson(lambda)
}
\arguments{
\item{lambda}{Optional fixed Poisson mean. If omitted, \code{lambda} is
computed from the model mean.}
}
\value{
A \code{Distribution} object.
}
\description{
The Poisson distribution.
}
