% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{Regular.polygon}
\alias{Regular.polygon}
\title{declare a regular polygon for the shape of a compartment in a flow char.}
\usage{
Regular.polygon(n)
}
\arguments{
\item{n}{the number of sides for the polygon.}
}
\value{
a character holding the tikz option for a node
}
\description{
declare a regular polygon for the shape of a compartment in a flow char.
}
