% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{class}
\name{Simulator}
\alias{Simulator}
\title{A R6 class that is the super class for all numerical simulators}
\description{
A R6 class that is the super class for all numerical simulators

A R6 class that is the super class for all numerical simulators
}
\details{
The main method is `simulate`.

the Simulation class is an abstract class. It's constructor
should not be explicitly called.

Functions used by the model must be explicited attached using the attach.function 
function provided by this package.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{a read-only field returning the  simulation code 
for the model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulator-new}{\code{Simulator$new()}}
\item \href{#method-Simulator-simulate}{\code{Simulator$simulate()}}
\item \href{#method-Simulator-clone}{\code{Simulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$new(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{the model to simulate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-simulate}{}}}
\subsection{Method \code{simulate()}}{
simulate the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$simulate(t, y0, parms = NULL, vars = names(y0), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{a numeric vector of times.}

\item{\code{y0}}{a numeric vector of initial conditions.}

\item{\code{parms}}{a numeric vector of parameter values.}

\item{\code{vars}}{a character vector specifying the variable names to be returned. 
A variable can be either a compartment or a substitution}

\item{\code{...}}{extra arguments to be passed to the `ode` method
substitutions that depend on the compartments are calculated 
and returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a data frame which rows correspond to each time in `t`, 
and columns correspond to the variables specified in vars.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
