% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{tikz.pdf}
\alias{tikz.pdf}
\title{generate a PDF file from the given latex commands}
\usage{
tikz.pdf(file, ..., fontsize = 12, preamble = NULL)
}
\arguments{
\item{file}{the path to the to-be-generated pdf file.}

\item{...}{the latex commands.}

\item{fontsize}{the fontsize in latex points.}

\item{preambles}{the latex preambles}
}
\description{
generate a PDF file from the given latex commands
}
\details{
the document class is "standalone", the tikz package is
automatically loaded.
}
\examples{
# an SIR model
SIR = Compartmental$new(S, I, R, title="SIR")
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
tikz.pdf("SIR.pdf", flowchart(SIR, TexFormatter$new(),
  S=Node(at=c(0,0)),
  I=Node(at=c(2,0)),
  R=Node(at=c(4,0))
))
}
