% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\name{Uniform}
\alias{Uniform}
\title{The uniform distribution}
\usage{
Uniform(min, max)
}
\arguments{
\item{min}{Optional minimum value.}

\item{max}{Optional maximum value.}
}
\value{
A \code{Distribution} object.
}
\description{
The uniform distribution.
}
