% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Beta}
\alias{Beta}
\title{The beta distribution}
\description{
The beta distribution

The beta distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{Beta}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Beta-new}{\code{Beta$new()}}
\item \href{#method-Beta-clone}{\code{Beta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$new(shape1, shape2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape1}}{the first shape parameter of the beta distribution}

\item{\code{shape2}}{the second shape parameter of the beta distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
