% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGillespie.R
\docType{class}
\name{CGillespie}
\alias{CGillespie}
\title{R6 class implementing the Gillespie method susing cpp11
 
This is a subclass of Simulator, implementing the Gillespie method
to simulate a compartmental model using cpp11. This class is only available
if the `cpp11` package is installed.}
\description{
R6 class implementing the Gillespie method susing cpp11
 
This is a subclass of Simulator, implementing the Gillespie method
to simulate a compartmental model using cpp11. This class is only available
if the `cpp11` package is installed.

R6 class implementing the Gillespie method susing cpp11
 
This is a subclass of Simulator, implementing the Gillespie method
to simulate a compartmental model using cpp11. This class is only available
if the `cpp11` package is installed.
}
\examples{
# an SIR model
SIR = Compartmental$new(S, I, R, title="SIR")
SIR$transition(S->I ~ beta*S*I/N, N=S+I+R, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
g = CGillespie$new(SIR)
g$simulate(0:100, y0=c(S=1000, I=10, R=0), parms=c(beta=0.4,gamma=0.2))
}
\section{Super class}{
\code{\link[REpiSim:Simulator]{REpiSim::Simulator}} -> \code{CGillespie}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CGillespie-new}{\code{CGillespie$new()}}
\item \href{#method-CGillespie-finalize}{\code{CGillespie$finalize()}}
\item \href{#method-CGillespie-clone}{\code{CGillespie$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Simulator" data-id="simulate"><a href='../../REpiSim/html/Simulator.html#method-Simulator-simulate'><code>REpiSim::Simulator$simulate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGillespie-new"></a>}}
\if{latex}{\out{\hypertarget{method-CGillespie-new}{}}}
\subsection{Method \code{new()}}{
constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGillespie$new(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{an object of the `Compartmental` class}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
the CGillespie class is only available if the cpp11 
package is installed.
# an SIR model
library(cpp11) # cpp11 is required to use CGillespie
SIR = Compartmental$new(S, I, R, N=S+I+R, title="SIR")
SIR$transition(S->I ~ beta*S*I/N, name="infection")
SIR$transition(I->R ~ gamma*I, name="recovery")
g = CGillespie$new(SIR)
g$simulate(0:100, y0=c(S=1000, I=10, R=0), parms=c(beta=0.4,gamma=0.2))
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGillespie-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-CGillespie-finalize}{}}}
\subsection{Method \code{finalize()}}{
the destructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGillespie$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The destructor unload the dynamic library loaded by cpp11
after compiling the C++ simulation code.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CGillespie-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CGillespie-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CGillespie$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
