% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Distribution}
\alias{Distribution}
\title{An R6 class that defines a distribution to be used as either a likelihood or a prior.}
\description{
An R6 class that defines a distribution to be used as either a likelihood or a prior.

An R6 class that defines a distribution to be used as either a likelihood or a prior.
}
\details{
Each Distribution object must provide two methods, a "log.density" method
which is the log-density function. In addition, a "log.likelihood" method which is
the log-likelihood function and the distirbution parameters
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{log.likelihood}}{The log-likelihood function, which takes as least two parameters:

x the observation to calculate the likelihood

mean the mean of the distribution. This typically corresponds to the model solution

... the parameters of the distribution,  in addition to the mean, to calculate the 
likelihood

It return the log likelihood}

\item{\code{log.density}}{the log-density function that takes a single paramter x,  which is
the value where the density is evaluated at.
it returns the log-density at x}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Distribution-new}{\code{Distribution$new()}}
\item \href{#method-Distribution-clone}{\code{Distribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Distribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-Distribution-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Distribution$new(..., .density)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the parameters needed to specify the distribution.}

\item{\code{.density}}{the density function}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
to use this object as a likelihood, additional parameters that are needed must
be provided as NA, while the other parameters that needs to be inferred from the mean
and the estimated parameters must be provided as a formula
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Distribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Distribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Distribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
