% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Exponential}
\alias{Exponential}
\title{The exponential distribution}
\description{
The exponential distribution

The exponential distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{Exponential}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Exponential-new}{\code{Exponential$new()}}
\item \href{#method-Exponential-clone}{\code{Exponential$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exponential-new"></a>}}
\if{latex}{\out{\hypertarget{method-Exponential-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exponential$new(rate)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rate}}{the rate parameter of the exponential distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Exponential-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Exponential-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exponential$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
