% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Gamma}
\alias{Gamma}
\title{The gamma distribution}
\description{
The gamma distribution

The gamma distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{Gamma}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Gamma-new}{\code{Gamma$new()}}
\item \href{#method-Gamma-clone}{\code{Gamma$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gamma-new"></a>}}
\if{latex}{\out{\hypertarget{method-Gamma-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gamma$new(shape, scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shape}}{the shape parameter of the gamma distribution}

\item{\code{scale}}{the scale parameter of the gamma distribution, i.e., 1/rate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gamma-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Gamma-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gamma$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
