% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TexFormatter.R
\name{merge.list}
\alias{merge.list}
\title{A S3 method to merge to lists}
\usage{
\method{merge}{list}(a, b)
}
\arguments{
\item{a}{a list}

\item{b}{a list}
}
\value{
a merged list
}
\description{
A S3 method to merge to lists
}
