% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{NBinom}
\alias{NBinom}
\title{The negative binomial distribution}
\description{
The negative binomial distribution

The negative binomial distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{NBinom}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NBinom-new}{\code{NBinom$new()}}
\item \href{#method-NBinom-clone}{\code{NBinom$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBinom-new"></a>}}
\if{latex}{\out{\hypertarget{method-NBinom-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBinom$new(size, prob)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{the size parameter of the negative binomial distribution}

\item{\code{prob}}{the probability parameter of the negative binomial distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBinom-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NBinom-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBinom$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
