% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{Normal}
\alias{Normal}
\title{The normal distribution.}
\description{
The normal distribution.

The normal distribution.
}
\details{
This distribution has one parameter named "size" in addition to the
mean, which is the size parameter in stats::dnbinom
}
\section{Super class}{
\code{REpiSim::Likelihood} -> \code{Normal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Normal-logL}{\code{Normal$logL()}}
\item \href{#method-Normal-clone}{\code{Normal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-logL"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-logL}{}}}
\subsection{Method \code{logL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$logL(x, mean, sd)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
