% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\docType{class}
\name{Normal}
\alias{Normal}
\title{The normal distribution}
\description{
The normal distribution

The normal distribution
}
\section{Super class}{
\code{\link[REpiSim:Likelihood]{REpiSim::Likelihood}} -> \code{Normal}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{par}}{return a vector of parameter names for this likelihood}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Normal-logL}{\code{Normal$logL()}}
\item \href{#method-Normal-clone}{\code{Normal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-logL"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-logL}{}}}
\subsection{Method \code{logL()}}{
The log likelihood function for the normal distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$logL(x, mean, sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the value to calculate the likelihood}

\item{\code{mean}}{the mean of the normal distribution}

\item{\code{sd}}{the standard deviation of the normal distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Normal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Normal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Normal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
