% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\docType{class}
\name{NormalProportional}
\alias{NormalProportional}
\title{The log likelihood function for the normal distribution}
\description{
A normal distribution which standard deviation is proportion to mean
}
\details{
This distribution has one parameter named "ceof" in addition to the
mean, where sd = coef*mean
}
\section{Super class}{
\code{\link[REpiSim:Likelihood]{REpiSim::Likelihood}} -> \code{NormalProportional}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{par}}{return a vector of parameter names for this likelihood}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NormalProportional-logL}{\code{NormalProportional$logL()}}
\item \href{#method-NormalProportional-clone}{\code{NormalProportional$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalProportional-logL"></a>}}
\if{latex}{\out{\hypertarget{method-NormalProportional-logL}{}}}
\subsection{Method \code{logL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalProportional$logL(x, mean, coef)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the value to calculate the likelihood}

\item{\code{mean}}{the mean of the normal distribution}

\item{\code{coef}}{the coefficient of variation, where sd = coef*mean}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalProportional-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NormalProportional-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalProportional$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
