% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizer.R
\name{Optimizer}
\alias{Optimizer}
\title{A subclass of Calibrator that uses an optimizer and an initial guess.
Calibrate the model to data}
\description{
A subclass of Calibrator that uses an optimizer and an initial guess.
Calibrate the model to data

A subclass of Calibrator that uses an optimizer and an initial guess.
Calibrate the model to data
}
\section{Super class}{
\code{REpiSim::Calibrator} -> \code{Optimizer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Optimizer-calibrate}{\code{Optimizer$calibrate()}}
\item \href{#method-Optimizer-clone}{\code{Optimizer$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="REpiSim" data-topic="Calibrator" data-id="initialize"><a href='../../REpiSim/html/Calibrator.html#method-Calibrator-initialize'><code>REpiSim::Calibrator$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-calibrate"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-calibrate}{}}}
\subsection{Method \code{calibrate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$calibrate(initial.values, parms, guess, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial.values}}{the initial values for the model. The parameters 
that need to be estimate should be NA, those that do not need to be
estimated must contain a finite value.}

\item{\code{parms}}{the parameter values of the model. The parameters 
that need to be estimate should be NA, those that do not need to be
estimated must contain a finite value.}

\item{\code{guess}}{the initial guess of the parameters to be fitted}

\item{\code{...}}{extra arguments to be passed to calibrators}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Optimizer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Optimizer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Optimizer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
