% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Poisson}
\alias{Poisson}
\title{The Poisson distribution}
\description{
The Poisson distribution

The Poisson distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{Poisson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Poisson-new}{\code{Poisson$new()}}
\item \href{#method-Poisson-clone}{\code{Poisson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poisson-new"></a>}}
\if{latex}{\out{\hypertarget{method-Poisson-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poisson$new(lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{the mean of the Poisson distribution. When getting the log-likelihood,
this parameter is unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poisson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Poisson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Poisson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
