% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{Star}
\alias{Star}
\title{declare a regular star for the shape of a compartment in a flow char.}
\usage{
Star(n, ratio = NULL)
}
\arguments{
\item{n}{the number of sides for the star}
}
\value{
a character holding the tikz option for a node
}
\description{
declare a regular star for the shape of a compartment in a flow char.
}
