% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distribution.R
\docType{class}
\name{Uniform}
\alias{Uniform}
\title{The uniform distribution}
\description{
The uniform distribution

The uniform distribution
}
\section{Super class}{
\code{\link[REpiSim:Distribution]{REpiSim::Distribution}} -> \code{Uniform}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Uniform-new}{\code{Uniform$new()}}
\item \href{#method-Uniform-clone}{\code{Uniform$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniform-new"></a>}}
\if{latex}{\out{\hypertarget{method-Uniform-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$new(min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min}}{the minimum value of the uniform distribution}

\item{\code{max}}{the maximum value of the uniform distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Uniform-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Uniform-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Uniform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
